gistax3/requirements.txt
requests
beautifulsoup4
PyPDF2
pymupdf
python-docx
tkinterdnd2
pdfplumber
docx-mailmerge
lxml